pyproject.toml
[project]
name = "skills"
version = "0.1.0"
description = "Reusable skills and agents for AI coding assistants"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "pyyaml>=6.0",
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"

[tool.ruff]
